Rent epoch fix

+++ src/main/java/org/p2p/solanaj/rpc/types/AccountInfo.java
@@ -21,7 +21,7 @@ public class AccountInfo extends RpcResultObject {
             this.executable = (boolean) am.get("executable");
             this.lamports = (long) (double) am.get("lamports");
             this.owner = (String) am.get("owner");
-            this.rentEpoch = (long) (double) am.get("rentEpoch");
+            this.rentEpoch = (double) am.get("rentEpoch");
         }
 
         @Json(name = "data")
@@ -37,7 +37,7 @@ public class AccountInfo extends RpcResultObject {
         private String owner;
 
         @Json(name = "rentEpoch")
-        private long rentEpoch;
+        private double rentEpoch;
     }
 
     @Json(name = "value")
